Move seed logic to management command
